refactor(webhook): inline hostname helpers and remove common package
The common package had a single caller, adding indirection without
reuse. Hostname validation and DNS lookup now live unexported next
to the validator that uses them.
